create a context for grid
