+++ src/Algorithms/dijkstra.js
@@ -4,6 +4,9 @@
 // by backtracking from the finish node.
 export function dijkstra(grid, startNode, finishNode) {
   console.log("grid rece : ",grid);
+  console.log("startNode : ",startNode);
+  console.log("finsihNode : ",finishNode);
+
   const visitedNodesInOrder = [];
   startNode.distance = 0;
   const unvisitedNodes = getAllNodes(grid);
